moved sponsor type to seperate type file
