+++ components/sponsors/GoldSponsors.tsx
@@ -12,9 +12,7 @@ interface GoldSponsorsProps {
  * @param {boolean} props.showSupportBanner - Indicates whether support banner should be displayed.
  */
 export default function GoldSponsors({
-  className = '',
-  // eslint-disable-next-line no-unused-vars, unused-imports/no-unused-vars
-  showSupportBanner = true
+  className = ''
 }: GoldSponsorsProps): React.ReactNode {
   return (
     <div className={`text-center ${className}`}>
