+++ types/components/sponsors/SponsorType.d.ts
